Move evaluated classical results under their own namespace
