+++ src/Parser/Evaluated/NaturalParser.php
@@ -8,17 +8,17 @@ namespace Serps\SearchEngine\Google\Parser\Evaluated;
 use Serps\SearchEngine\Google\Page\GoogleDom;
 use Serps\SearchEngine\Google\Parser\AbstractParser;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\AnswerBox;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ClassicalResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Divider;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Flight;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\InTheNews;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\LargeClassicalResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\LargeClassicalResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Map;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\NewDesignClassicalResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarousel;
-use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ClassicalWithLargeVideo;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalWithLargeVideo;
 
 /**
  * Parses natural results from a google SERP
@@ -35,7 +35,7 @@ class NaturalParser extends AbstractParser
             new Divider(),
             new SearchResultGroup(),
             new ClassicalResult(),
-            new NewDesignClassicalResult(),
+            new ClassicalCardsResult(),
             new LargeClassicalResult(),
             new ImageGroup(),
             new TweetsCarousel(),
