+++ test/suites/Parser/Evaluated/NaturalParserTest.php
@@ -24,17 +24,18 @@ use Symfony\Component\Yaml\Yaml;
  * For instance if the previous test included a ``TweetsCarousel`` the new test should do so.
  *
  *
- * @covers Serps\SearchEngine\Google\Parser\AbstractParser
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\NaturalParser
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ClassicalResult
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\LargeClassicalResult
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarousel
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\InTheNews
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Divider
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ClassicalWithLargeVideo
- * @covers Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Map
+ * @covers \Serps\SearchEngine\Google\Parser\AbstractParser
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\NaturalParser
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalResult
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\LargeClassicalResult
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalWithLargeVideo
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResult
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarousel
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\InTheNews
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Divider
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup
+ * @covers \Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Map
  */
 class NaturalParserTest extends GoogleSerpTestCase
 {
